Section 6 finished.
It was really a hard work to complete section 6, which is also an excuse for delaying this translation... and the changes in format is due to my new VS2017 editor. Pardon me. Now I'm 32 sections in debts, and I'm going to translate one section per week since my summer vacation is ending. That's all I want to say.